renamed relay test app
