+++ CursorParty/.gitignore
