+++ CursorParty/src/com/bitheads/braincloud/GameScreen.java
@@ -1,4 +1,4 @@
-package com.bitheads.relaytestapp;
+package com.bitheads.braincloud;
 
 import java.awt.Color;
 import java.awt.Dimension;
