+++ CursorParty/src/com/bitheads/braincloud/Lobby.java
@@ -1,4 +1,4 @@
-package com.bitheads.relaytestapp;
+package com.bitheads.braincloud;
 
 import java.util.ArrayList;
