+++ CursorParty/src/com/bitheads/braincloud/MainMenuScreen.java
@@ -1,4 +1,4 @@
-package com.bitheads.relaytestapp;
+package com.bitheads.braincloud;
 
 import java.awt.Dimension;
 import java.awt.Font;
