+++ CursorParty/src/com/bitheads/braincloud/Screen.java
@@ -1,4 +1,4 @@
-package com.bitheads.relaytestapp;
+package com.bitheads.braincloud;
 
 import javax.swing.JPanel;
