+++ CursorParty/src/com/bitheads/braincloud/User.java
@@ -1,4 +1,4 @@
-package com.bitheads.relaytestapp;
+package com.bitheads.braincloud;
 
 import java.awt.geom.Point2D;
