+++ CursorParty/src/resources/arrow0.png
