+++ CursorParty/src/resources/arrow3.png
