+++ CursorParty/src/resources/arrow4.png
