+++ CursorParty/src/resources/arrow5.png
