+++ CursorParty/src/resources/arrow6.png
