+++ CursorParty/src/resources/arrow7.png
