--- RelayTestApp/src/com/bitheads/relaytestapp/Version.java
@@ -1,6 +1,0 @@
-package com.bitheads.relaytestapp;
-
-class Version
-{
-    static public String version = "5.4.0";
-}
